compute_modules/image_processor/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image_processor"
version = "0.1.0"
description = "Image Processor Compute Module"
requires-python = ">=3.9"
dependencies = [
    "pillow>=10.0.0",
    "sqlalchemy>=2.0.0",
    "paho-mqtt>=1.6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.compute_module]
supported_tasks = ["image_resize", "image_conversion"]
